removing job calculation logic from prepare_locations.py
